Fama and French portfolio daily returns and factors for international markets.

A dataset containing the returns of four portfolios ordered by size and book-to-market. The four portfolios are SMALL/LoBM, SMALL/HiBM, BIG/LoBM and BIG/HiBM in four international markets: North America (NA), Japan (JP), Asia Pacific (AP) and Europe (EU). It also contains the Fama/French 5 factors for each of the markets.

Format

A data frame with 314 rows and 41 variables.

Date

Date, months from July 1990 until August 2016

NA.SMALL.LoBM

Monthly returns of portfolio SMALL/LoBM in North American market

NA.SMALL.HiBM

Monthly returns of portfolio SMALL/HiBM in North American market

NA.BIG.LoBM

Monthly returns of portfolio BIG/LoBM in North American market

NA.BIG.HiBM

Monthly returns of portfolio BIG/HiBM in North American market

NA.Mkt.RF

North American market excess returns, i.e return of the market - market risk free rate

NA.SMB

SMB (Small Minus Big) for the North American market

NA.HML

HML (High Minus Low) for the North American market

NA.RMW

RMW (Robust Minus Weak) for the North American market

NA.CMA

CMA (Conservative Minus Aggressive) for the North American market

NA.RF

North American risk free rate

JP.SMALL.LoBM

Monthly returns of portfolio SMALL/LoBM in Japanese market

JP.SMALL.HiBM

Monthly returns of portfolio SMALL/HiBM in Japanese market

JP.BIG.LoBM

Monthly returns of portfolio BIG/LoBM in Japanese market

JP.BIG.HiBM

Monthly returns of portfolio BIG/HiBM in Japanese market

JP.Mkt.RF

Japanese market excess returns, i.e return of the market - market risk free rate

JP.SMB

SMB (Small Minus Big) for the Japanese market

JP.HML

HML (High Minus Low) for the Japanese market

JP.RMW

RMW (Robust Minus Weak) for the Japanese market

JP.CMA

CMA (Conservative Minus Aggressive) for the Japanese market

JP.RF

Japanese risk free rate

AP.SMALL.LoBM

Monthly returns of portfolio SMALL/LoBM in Asia Pacific market

AP.SMALL.HiBM

Monthly returns of portfolio SMALL/HiBM in Asia Pacific market

AP.BIG.LoBM

Monthly returns of portfolio BIG/LoBM in Asia Pacific market

AP.BIG.HiBM

Monthly returns of portfolio BIG/HiBM in Asia Pacific market

AP.Mkt.RF

Asia Pacific market excess returns, i.e return of the market - maket risk free rate

AP.SMB

SMB (Small Minus Big) for the Asia Pacific market

AP.HML

HML (High Minus Low) for the Asia Pacific market

AP.RMW

RMW (Robust Minus Weak) for the Asia Pacific market

AP.CMA

CMA (Conservative Minus Aggressive) for the Asia Pacific market

AP.RF

Asia Pacific risk free rate

EU.SMALL.LoBM

Excess return of portfolio SMALL/LoBM in European market

EU.SMALL.HiBM

Excess return of portfolio SMALL/HiBM in European market

EU.BIG.LoBM

Excess return of portfolio BIG/LoBM in European market

EU.BIG.HiBM

Excess return of portfolio BIG/HiBM in European market

EU.Mkt.RF

European market excess returns, i.e returns of the market - market risk free rate

EU.SMB

SMB (Small Minus Big) for the European market

EU.HML

HML (High Minus Low) for the European market

EU.RMW

RMW (Robust Minus Weak) for the European market

EU.CMA

CMA (Conservative Minus Aggressive) for the European market

EU.RF

European risk free rate

Source

http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html

References

Kennet R. French - Data Library (2017) http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html#International

Fama, E. and French, K. R (1993) Common risk factors in the returns on stocks and bonds, Journal of Financial Economics, 3-56.

Fama, E. F. and French, K. R (2015) A five-factor asset pricing model, Journal of Financial Economics, 116, 1-22.