Forecast Methods for Objects in tvReg.

Forecast methods for objects with class attribute tvlm, tvar, tvvar, tvirf and tvsure. If the smoothing variable (z) in the model is non-NULL and it is a random variable then use function predict with parameter newz.

forecast(object, ...)

# S3 method for tvlm
forecast(object, newx, n.ahead = 1, winsize = 0, ...)

# S3 method for tvar
forecast(object, n.ahead = 1, newexogen = NULL,
  winsize = 0, ...)

# S3 method for tvvar
forecast(object, n.ahead = 1, newexogen = NULL,
  winsize = 0, ...)

# S3 method for tvsure
forecast(object, newdata, n.ahead = 1, winsize = 0,
  ...)

Arguments

object
Object of class tvlm, tvar, tvvar or tvsure.

...
Other parameters passed to specific methods.

newx
A vector, dataframe or matrix with new values of all variables in x. No need to input the intercept.

n.ahead
A scalar with the forecast horizon, value 1 by default.

winsize
A scalar. If 0 then an 'increase window' forecasting is performed. Otherwise a 'rolling window' forecasting is performed with window size given by 'winsize'.

newexogen
A matrix or vector with the new value of the exogenous variables. Only for predictions of *tvar* and *tvvar* objects.

newdata
A matrix or data.frame with the values of the regressors to use for forecasting.

Value

An object of class matrix or vector with the same dimensions than the dependent variable of object.

See also

Examples

data("RV")
RV2 <- head(RV, 2001)
tvHAR <- tvLM (RV ~ RV_lag + RV_week + RV_month, data = RV2, bw = 20)
newx <- cbind(RV$RV_lag[2002:2004], RV$RV_week[2002:2004],
              RV$RV_month[2002:2004])
forecast(tvHAR, newx, n.ahead = 3)

#> [1] 2.529025e-05 2.223299e-05 2.267907e-05

exogen = RV2[, c("RV_week", "RV_month")]
tvHAR2 <- tvAR(RV2$RV_lag, p = 1, exogen = exogen, bw = 20)
newexogen <- newx[, -1]
forecast(tvHAR2, n.ahead = 3, newexogen = newexogen)

#> [1] 2.388977e-05 1.811485e-05 1.449888e-05

data(usmacro, package = "bvarsv")
tvVAR <- tvVAR(usmacro, p = 6, type = "const", bw = c(1.8, 20, 20))
forecast(tvVAR, n.ahead = 10)

#>            inf      une      tbi
#>  [1,] 2.295453 5.130214 2.985555
#>  [2,] 2.281132 5.192859 2.874300
#>  [3,] 2.224442 5.110938 2.528473
#>  [4,] 2.124246 4.927610 2.312528
#>  [5,] 2.112547 4.767046 2.585776
#>  [6,] 2.174060 4.659349 2.830172
#>  [7,] 2.299620 4.606723 2.887135
#>  [8,] 2.496606 4.569793 2.961808
#>  [9,] 2.702840 4.534813 3.098251
#> [10,] 2.913992 4.512365 3.367365

data("Kmenta", package = "systemfit")
eqDemand <- consump ~ price + income
eqSupply <- consump ~ price + farmPrice
system <- list(demand = eqDemand, supply = eqSupply)
tvOLS.fit <- tvSURE(system, data = Kmenta, est = "ll", bw = c(1.5, 1.5))
newdata <- data.frame(consump = c(95, 100, 102), price = c(90, 100, 103),
farmPrice = c(70, 95, 103), income = c(82, 94, 115))
forecast(tvOLS.fit, newdata = newdata, n.ahead = 3)

#>         demand      supply
#> [1,]  94.37623 -15.2166191
#> [2,]  97.91204  -2.2374701
#> [3,] 104.29284   0.1018483