Plot Methods for Objects in tvReg

Plot methods for objects with class attribute tvlm, tvar, tvvar, tvirf, tvsure or tvplm.

# S3 method for tvsure
plot(x, eqs = NULL, vars = NULL, plot.type = c("multiple", "single"), ...)

# S3 method for tvlm
plot(x, ...)

# S3 method for tvar
plot(x, ...)

# S3 method for tvplm
plot(x, ...)

# S3 method for tvvar
plot(x, ...)

# S3 method for tvirf
plot(
  x,
  obs.index = NULL,
  impulse = NULL,
  response = NULL,
  plot.type = c("multiple", "single"),
  ...
)

Arguments

x
An object used to select the method.

eqs
A vector of integers. Equation(s) number(s) of the coefficients to be plotted.

vars
A vector of integers. Variable number(s) of the coefficients to be plotted.

plot.type
Character, if multiple all plots are drawn in a single device, otherwise the plots are shown consecutively.

...
Other parameters passed to specific methods.

obs.index
Scalar (optional), the time at which the impulse response is plotted. If left NULL, the mean over the whole period is plotted (this values should be similar to the estimation using a non time-varying VAR method).

impulse
Character vector (optional) of the impulses, default is all variables.

response
Character vector (optional) of the responses, default is all variables.

See also