tvReg: Time-Varying Coefficient for Single and Multi-Equation Regressions

This package covers a large range of semiparametric regression methods with time-varying coefficients using nonparametric kernel smoothing for the estimation.

Main functions

The five basic functions in this package are tvLM, tvAR, tvSURE, tvPLM, tvVAR and tvIRF. Moreover, this package provides the confint, fitted, forecast, plot, predict, print, resid and summary methods adapted to the class attributes of the tvReg. In addition, it includes bandwidth selection methods, time-varying variance-covariance estimators and four estimation procedures: the time-varying ordinary least squares, which are implemented in the tvOLS methods, the time-varying generalised least squares for a list of equations, which is implemented in the tvGLS methods, time-varying pooled and random effects estimators for panel data, which are implemented in the tvRE and the time-varying fixed effects estimator, which is implemente in the tvFE.

Further information

Details on the theory and applications to finance and macroeconomics can be found in Casas and Fernandez-Casal (2019, https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3363526), and in the package vignette https://icasas.github.io/tvReg/articles/tvReg.html.

Acknowledgments

Funded by the Horizon 2020. Framework Programme of the European Union.

References

Casas, I. and Fernandez-Casal, R., tvReg: Time-varying Coefficient Linear Regression for Single and Multi-Equations in R (April 1, 2019). Available at SSRN: https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3363526.

Author

Isabel Casas (casasis@gmail.com), Ruben Fernandez-Casal (rubenfcasal@gmail.com).